+++ lib/ig_markets/deal_confirmation.rb
@@ -14,7 +14,7 @@ module IGMarkets
     attribute :deal_status, Symbol, allowed_values: [:accepted, :fund_account, :rejected]
     attribute :direction, Symbol, allowed_values: [:buy, :sell]
     attribute :epic
-    attribute :expiry, Date, nil_if: %w(- DFB), format: '%d-%b-%y'
+    attribute :expiry, Date, nil_if: %w(- DFB), format: ['%d-%b-%y', '%b-%y']
     attribute :guaranteed_stop, Boolean
     attribute :level, Float
     attribute :limit_distance, Fixnum