+++ spec/ig_markets/instrument_spec.rb
@@ -1,6 +1,9 @@
 describe IGMarkets::Instrument do
-  it 'accepts DFB as an expiry' do
+  it 'accepts different expiry formats' do
+    expect(build(:instrument, expiry: '-').expiry).to eq(nil)
     expect(build(:instrument, expiry: 'DFB').expiry).to eq(nil)
+    expect(build(:instrument, expiry: 'SEP-16').expiry).to eq(Date.new(2016, 9, 1))
+    expect(build(:instrument, expiry: '08-DEC-16').expiry).to eq(Date.new(2016, 12, 8))
   end
 
   it 'adjusts API attributes' do